make python.sh idempotent

+++ install-root/python.sh
@@ -17,7 +17,7 @@ apt-get install -y --no-install-recommends python3.7 python3.7-dev python3.7-ven
 [[ ! -f /usr/local/bin/python ]] && ln -s /usr/bin/python3.7 /usr/local/bin/python
 
 # allow apt_pkg to be used by python minor versions other than the python3-apt build version
-ln -s /usr/lib/python3/dist-packages/apt_pkg.cpython-*.so /usr/lib/python3/dist-packages/apt_pkg.so
+[[ ! -f /usr/lib/python3/dist-packages/apt_pkg.so ]] && ln -s /usr/lib/python3/dist-packages/apt_pkg.cpython-*.so /usr/lib/python3/dist-packages/apt_pkg.so
 
 # install pip directly, rather than installing the deb package which depends on the older python3 package
 # use sudo to make sure it is installed as a system rather than user package (ie: /usr/local/lib/python3.7/dist-packages/)
